feat :sparkles: Depot argent features

+++ banque/Transaction.cs
@@ -1,10 +1,10 @@
 namespace banque {
     public class Transaction{
         public string intitule{get; set;}
-        public int montant{get; set;}
+        public double montant{get; set;}
         public DateTime date {get; set;}
 
-        public Transaction(string intituleTrans,int montantTrans){
+        public Transaction(string intituleTrans,double montantTrans){
             if(intitule != ""){
                 intitule=intituleTrans;
             }else{
